+++ gui/src/main/daemon-rpc.ts
@@ -259,17 +259,38 @@ const tunnelStateSchema = oneOf(
   object({
     state: enumeration('error'),
     details: object({
-      is_blocking: boolean,
+      block_failure: maybe(
+        object({
+          reason: enumeration('generic', 'locked'),
+          details: maybe(
+            object({
+              name: string,
+              pid: number,
+            }),
+          ),
+        }),
+      ),
       cause: oneOf(
         object({
           reason: enumeration(
             'ipv6_unavailable',
-            'set_firewall_policy_error',
             'set_dns_error',
             'start_tunnel_error',
             'is_offline',
             'tap_adapter_problem',
           ),
+        }),
+        object({
+          reason: enumeration('set_firewall_policy_error'),
+          details: object({
+            reason: enumeration('generic', 'locked'),
+            details: maybe(
+              object({
+                name: string,
+                pid: number,
+              }),
+            ),
+          }),
         }),
         object({
           reason: enumeration('auth_failed'),